requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
email-validator==2.1.1
WTForms==3.1.2
SQLAlchemy==2.0.40
Jinja2==3.1.6
Werkzeug==3.1.3
click==8.1.8
itsdangerous==2.2.0
MarkupSafe==3.0.2
blinker==1.9.0
python-dotenv==1.0.1
